insurance-extraction-engine/services/normalizer/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dateutil==2.8.2
loguru==0.7.2